fix: update styles and file
