+++ src/index.ts
@@ -1,15 +1,3 @@
-import validateValueOnInput from './js/testingFin';
 import './sass/styles.scss';
-// import './assets/horse.svg';
-// import "./style.css";
 
-console.log('Hello my team');
-
-const img: HTMLImageElement = document.createElement('img');
-img.setAttribute('src', './img/horse.svg');
-document.body.appendChild(img);
-
-const input = document.createElement('input');
-document.body.appendChild(input);
-
-input.addEventListener('input', () => validateValueOnInput(input));
+import './js/testingFin';
