+++ src/sass/abstract/_constants.scss
@@ -1,5 +1,19 @@
-$color-body: #d4f132;
-$color-auto-text: #ee17bf;
+$color-light: #f5f5f5;
+$color-primary: #ff5a00;
+// $color-additional: #422779;
+$color-additional: #650e70;
+
+$color-primary-light: #ff8e08;
+
+$color-text: #f5f5f5;
+$color-dark: #303030;
+$color-border: #3a4e58;
+$color-link: #422779;
+$color-hovered: #ff5a00;
+
+$color-body: #ff8e08;
+
+$color-overlay: rgba(80, 80, 80, 0.4);
 
 $content-width: 1360px;
 $desktop-small-width: 1024px;
