+++ src/sass/abstract/_reset.scss
@@ -16,5 +16,5 @@ li {
 
 a {
     text-decoration: none;
-    color: $color-auto-text;
+    color: $color-link;
 }
