+++ src/sass/base/_buttons.scss
@@ -1,0 +1,3 @@
+.button-basic {
+  @include button-basic;
+}
