+++ src/sass/pages/_login.scss
@@ -1,0 +1,3 @@
+.input_login {
+    @include input;
+}
